Get value index

Returns the (not chain-linked) series of value indices from a series of current prices.

Usage

get_v_index(current)

Arguments

current

(ts) Series of current prices series.

Value

(ts) Series of value indices.

Details

The value of the resulting series x at (y,s), where y is the year and s is the subyear period, is current(y,s)/current(y)

Examples

get_v_index(gdp_current)
#>           Qtr1      Qtr2      Qtr3      Qtr4
#> 1996 100.51732 107.13707 104.36124 112.91555
#> 1997  99.66009 106.72917 103.94849 114.22529
#> 1998 100.44954 107.87424 105.07408 114.80256
#> 1999 101.42289 107.82941 104.98399 114.33529
#> 2000 103.24540 109.44658 106.78413 115.68740
#> 2001 102.82951 109.02745 106.22436 114.89679
#> 2002 101.43033 108.34829 104.74979 113.31121
#> 2003 101.86997 108.25028 104.89020 113.23332
#> 2004 101.59976 108.29281 104.95002 113.74951
#> 2005 102.23950 109.41947 105.51788 114.47819
#> 2006 102.80416 109.66360 105.73308 114.92153
#> 2007 102.15368 108.75961 104.05462 113.91436
#> 2008  99.66544 106.12534 100.43683 106.72448
#> 2009  93.07823  98.19998  94.03649 100.50304
#> 2010  96.36511 102.39499  97.96811 104.82073
#> 2011  96.49917 101.68882  96.79625 101.87598
#> 2012  94.85632  98.74669  95.25456  98.89341
#> 2013  95.77437 100.63786  97.50146 102.10471
#> 2014  97.33321 102.36025 100.02184 105.47047
#> 2015  99.85437 105.72300 103.48227 109.48333
#> 2016  98.68680 104.56991 102.25662 107.67943
#> 2017  99.20149 105.51192 103.08299 108.96527
#> 2018  98.65370 105.00708 102.28166 108.50239
#> 2019  99.37605 104.86127 101.63758 107.79657
#> 2020  92.77393  80.44859  90.57597  96.48069
#> 2021 100.94491 108.59979 108.82047 119.27518
#> 2022 103.64314 111.26839 110.64029 119.17758
#> 2023 103.88016 109.46129 107.63489 115.33478
#> 2024 100.87698 106.92427 105.14548 112.36475